Move Transaction stateful validations into submodule
